refactor(website): split the whitepaper markdown helpers out so they can be tested
`nx test website` failed where `vitest --root apps/website` passed, and nx
swallowed the output entirely — the sign of a hard crash during collection
rather than a failing assertion. The cause was the spec importing
generate-whitepaper.ts, which pulls in puppeteer and the Anthropic SDK at
module scope.

That is also the reason these functions had no spec in the first place, and
therefore the reason three text-DELETING defects reached published PDFs
unnoticed. A pure string function should not drag a browser-automation library
into a jsdom worker, so mdToHTML and escapeHtml now live in
whitepaper-markdown.ts and the generator imports them.

Verified the spec genuinely runs rather than being skipped: 137 files / 1382
tests under `nx test website`, with the 7 markdown cases among them. The
generator still produces byte-identical HTML.
